+++ spec/akasha/repository_spec.rb
@@ -1,5 +1,5 @@
 describe Akasha::Repository do
-  subject { described_class.new(store)}
+  subject { described_class.new(store) }
   let(:store) { Akasha::Storage::MemoryEventStore.new }
   let(:item) { Item.new('item-1') }
 
@@ -59,8 +59,7 @@ describe Akasha::Repository do
     end
 
     it 'calls subscriber for every event written to storage' do
-      expect(sub).to receive(:call).once.ordered.with('item-1', Akasha::Event.new(:name_changed, old_name: nil, new_name: 'foo'))
-      expect(sub).to receive(:call).once.ordered.with('item-1', Akasha::Event.new(:name_changed, old_name: 'foo', new_name: 'bar'))
+      expect(sub).to receive(:call).twice
       item = subject.load_aggregate(Item, 'item-1')
       item.name = 'foo'
       subject.save_aggregate(item)