+++ spec/fixtures/item_logger.rb
@@ -1,7 +1,6 @@
 # An example event listener.
-class Logger < Akasha::EventListener
+class ItemLogger < Akasha::EventListener
   # Handle the 'name_changed' event.
   # Specs will spy on this method.
-  def on_name_changed(_aggregate_id, **)
-  end
+  def on_name_changed(_aggregate_id, **); end
 end