+++ spec/fixtures/notifier.rb
@@ -2,6 +2,5 @@
 class Notifier < Akasha::EventListener
   # Handle the 'name_changed' event.
   # Specs will spy on this method.
-  def on_name_changed(_aggregate_id, new_name:, **)
-  end
+  def on_name_changed(_aggregate_id, new_name:, **); end
 end